light cleanup, fix titles
